[Q20-15] Worked on review comments regarding Custom mappings, separated business logic in MapGraphToTripleController.
