--- quadriga/src/main/java/edu/asu/diging/quadriga/core/service/MapGraphToTriple.java
@@ -1,5 +1,0 @@
-package edu.asu.diging.quadriga.core.service;
-
-public interface MapGraphToTriple {
-
-}
